Create README.md
PRESENTATION
The purpose of Stockholm Culture Awards is to highlight and praise world-class personalities
within culture, which are knowledgeable and distinguished in their field and work hard to
strengthen the status of classical arts in and for Stockholm.
Registered in Sweden.
Registration number 802461–7030.
About Stockholm Culture Awards:
The Stockholm Culture Foundation was formed in 1983 by a group of students, who became
acquainted with each other during the university studies at Uppsala University. Most of them
were active members of Stockholm’s Nation, in the latter part of the 1970s and early 1980s.
The common interest was Stockholm's culture and history.
Stockholm Culture Awards thrive in making Stockholm a leading hub of knowledge and art.
Stockholm Culture Awards want to make their awards more accessible to all and both
Swedish and foreign citizens are eligible.
Stockholm Culture Awards gather members and others interested in and for Stockholm’s
culture, history and development in a rapidly changing world. Over the years, the circle of
interested people has expanded.
Over the years, the circle of interested people in Stockholm Culture Awards has expanded and
the focus on the city’s culture and history, in an international context has grown.
Stockholm Culture Awards is also interested in the impact from other countries on the city’s
cultural and historical development, with the purpose to strengthen Stockholm in its role as a
city of culture and knowledge.
The Stockholm Culture Foundation brings together its stakeholders to an annual awarding of
medals and diplomas, under the name Stockholm Culture Awards, with a subsequent dinner
and dancing.
STOCKHOLM CULTURE AWARDS
 Page 2
Stockholm Culture Awards grant Merit Medals and Diplomas in the
following categories:
Classic and Modern Music Classic Art
Family History Film
Architecture Dance
Cultural Mathematics The art of food as culture
Theater Sculpture
Opera Literature
STOCKHOLM CULTURE AWARDS